test-requirements.txt
# Test requirements
beautifulsoup4
Pillow
pytest-django
